Query Json Object & Refactors
