+++ src/main/java/com/example/studentservice/controller/DisplayRestController.java
@@ -2,9 +2,10 @@ package com.example.studentservice.controller;
 
 import com.example.studentservice.domain.student.Student;
 import com.example.studentservice.dto.ClassInfoDTO;
-import com.example.studentservice.dto.GradeQuery;
+import com.example.studentservice.query.GradeQuery;
 import com.example.studentservice.dto.StudentDTO;
 import com.example.studentservice.dto.StudentGradeDTO;
+import com.example.studentservice.query.StudentQuery;
 import com.example.studentservice.service.ClassInfoService;
 import com.example.studentservice.service.GradeService;
 import com.example.studentservice.service.StudentService;
@@ -39,8 +40,8 @@ public class DisplayRestController {
 
     @RequestMapping("/student")
     @CrossOrigin
-    public StudentDTO findStudent(@RequestBody String sName, @RequestBody String sSerial) {
-        return studentService.getInfo(sName, sSerial);
+    public StudentDTO findStudent(@RequestBody StudentQuery studentQuery) {
+        return studentService.getInfo(studentQuery.getStudentSerial(), studentQuery.getStudentName());
     }
 
     @RequestMapping("/classInfo")
