+++ src/main/java/com/example/studentservice/query/StudentQuery.java
@@ -1,0 +1,13 @@
+package com.example.studentservice.query;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class StudentQuery {
+    String studentName;
+    String studentSerial;
+}
